changes to ensure wellbeing test pass

+++ api/__test__/mock/database/setup.sql
@@ -31,7 +31,7 @@ CREATE TABLE plant (
     nickname VARCHAR(200) NOT NULL,
     name VARCHAR(200) NOT NULL,
     trefle_id INT,
-    wellbeing_rating INT DEFAULT 100 CHECK (
+    wellbeing_rating FLOAT DEFAULT 100 CHECK (
         wellbeing_rating BETWEEN 0
         AND 100
     ),
